fix: context 파일 분리
