+++ src/app/relationreport/[reportSeq]/page.tsx
@@ -5,60 +5,8 @@ import { InviteFriendsPopup, RelationReportPageContainer } from './components';
 import AddFriendsPopup from './components/popup/AddFriendsPopup';
 import EditMoimPopup from './components/popup/EditMoimPopup';
 import PreventSharePopup from './components/popup/PreventSharePopup';
-import { useParams, useSearchParams } from 'next/navigation';
 import { Node } from 'reactflow';
-
-interface IRelationReportModalContext {
-  isInviteFriendsPopupOpen: boolean;
-  setInviteFriendsPopupOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  isAddFriendsPopupOpen: boolean;
-  setIsAddFriendsPopupOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  editMoimPopupInfo: {
-    title: string;
-    isPrivate: boolean;
-  } | null;
-  setEditMoimPopupInfo: React.Dispatch<
-    React.SetStateAction<{
-      title: string;
-      isPrivate: boolean;
-    } | null>
-  >;
-  isPreventSharePopupOpen: boolean;
-  setIsPreventSharePopupOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  isAllLoading: boolean;
-  setIsAllLoading: React.Dispatch<React.SetStateAction<boolean>>;
-  isOnlyEdge: boolean; // 관계도에 엣지가 한 개 인가?
-  setIsOnlyEdge: React.Dispatch<React.SetStateAction<boolean>>;
-  playData?: any[];
-  setPlayData: React.Dispatch<React.SetStateAction<any[]>>;
-  initialNodes: Node<any, string | undefined>[]; // 그래프의 초기 위치로 뒤돌리기위한 첫 노드 정보
-  setInitialNodes: (
-    value: React.SetStateAction<Node<any, string | undefined>[]>,
-  ) => void;
-}
-
-const initialState = {
-  isInviteFriendsPopupOpen: false,
-  setInviteFriendsPopupOpen: () => {},
-  isAddFriendsPopupOpen: false,
-  setIsAddFriendsPopupOpen: () => {},
-  isPreventSharePopupOpen: false,
-  setIsPreventSharePopupOpen: () => {},
-  editMoimPopupInfo: null,
-  setEditMoimPopupInfo: () => {},
-  isAllLoading: false,
-  setIsAllLoading: () => {},
-  isOnlyEdge: false,
-  setIsOnlyEdge: () => {},
-  setOnAddNodeClick: () => {},
-  playData: [],
-  setPlayData: () => {},
-  initialNodes: [],
-  setInitialNodes: () => {},
-};
-
-export const RelationReportModalContext =
-  React.createContext<IRelationReportModalContext>(initialState);
+import RelationReportContext from './RelationReportContext';
 
 const RelationReportPage = () => {
   const [isInviteFriendsPopupOpen, setInviteFriendsPopupOpen] =
@@ -77,9 +25,11 @@ const RelationReportPage = () => {
   const [initialNodes, setInitialNodes] = React.useState<
     Node<any, string | undefined>[]
   >([]);
+  console.log(process.env.NODE_ENV);
+  // yarn dev : development
 
   return (
-    <RelationReportModalContext.Provider
+    <RelationReportContext.Provider
       value={{
         isInviteFriendsPopupOpen,
         setInviteFriendsPopupOpen,
@@ -116,7 +66,7 @@ const RelationReportPage = () => {
       <AddFriendsPopup onClose={() => setIsAddFriendsPopupOpen(false)} />
       <PreventSharePopup onClose={() => setIsPreventSharePopupOpen(false)} />
       {editMoimPopupInfo && <EditMoimPopup />}
-    </RelationReportModalContext.Provider>
+    </RelationReportContext.Provider>
   );
 };
